Struct rusoto_events::CloudWatchEventsClient

pub struct CloudWatchEventsClient { /* fields omitted */ }

A client for the Amazon CloudWatch Events API.

Methods

impl CloudWatchEventsClient

pub fn new(region: Region) -> CloudWatchEventsClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> CloudWatchEventsClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,