Struct rusoto_events::PutPermissionRequest

pub struct PutPermissionRequest {
    pub action: String,
    pub principal: String,
    pub statement_id: String,
}

Fields
action: String

The action that you are enabling the other account to perform. Currently, this must be events:PutEvents.

principal: String

The 12-digit AWS account ID that you are permitting to put events to your default event bus. Specify "*" to permit any account to put events to your default event bus.

If you specify "*", avoid creating rules that may match undesirable events. To create more secure rules, make sure that the event pattern for each rule contains an account field with a specific account ID from which to receive events. Rules with an account field do not match any events sent from other accounts.

statement_id: String

An identifier string for the external account that you are granting permissions to. If you later want to revoke the permission for this external account, specify this StatementId when you run RemovePermission.