Crate rusoto_events

Amazon CloudWatch Events helps you to respond to state changes in your AWS resources. When your resources change state, they automatically send events into an event stream. You can create rules that match selected events in the stream and route them to targets to take action. You can also use rules to take action on a predetermined schedule. For example, you can configure rules to:

Automatically invoke an AWS Lambda function to update DNS entries when an event notifies you that Amazon EC2 instance enters the running state.

Direct specific API records from AWS CloudTrail to an Amazon Kinesis data stream for detailed analysis of potential security or availability risks.

Periodically invoke a built-in target to create a snapshot of an Amazon EBS volume.

For more information about the features of Amazon CloudWatch Events, see the Amazon CloudWatch Events User Guide.

If you're using the service, you're probably looking for CloudWatchEventsClient and CloudWatchEvents.

Structs

AwsVpcConfiguration
This structure specifies the VPC subnets and security groups for the task, and whether a public IP address is to be used. This structure is relevant only for ECS tasks that use the awsvpc network mode.

BatchArrayProperties
The array properties for the submitted job, such as the size of the array. The array size can be between 2 and 10,000. If you specify array properties for a job, it becomes an array job. This parameter is used only if the target is an AWS Batch job.

BatchParameters
The custom parameters to be used when the target is an AWS Batch job.

BatchRetryStrategy
The retry strategy to use for failed jobs, if the target is an AWS Batch job. If you specify a retry strategy here, it overrides the retry strategy defined in the job definition.

CloudWatchEventsClient
A client for the Amazon CloudWatch Events API.

Condition
A JSON string which you can use to limit the event bus permissions you are granting to only accounts that fulfill the condition. Currently, the only supported condition is membership in a certain AWS organization. The string must contain Type, Key, and Value fields. The Value field specifies the ID of the AWS organization. Following is an example value for Condition:

'{"Type" : "StringEquals", "Key": "aws:PrincipalOrgID", "Value": "o-1234567890"}'

DeleteRuleRequest

DescribeEventBusRequest

DescribeEventBusResponse

DescribeRuleRequest

DescribeRuleResponse

DisableRuleRequest

EcsParameters
The custom parameters to be used when the target is an Amazon ECS task.

EnableRuleRequest

InputTransformer
Contains the parameters needed for you to provide custom input to a target based on one or more pieces of data extracted from the event.

KinesisParameters
This object enables you to specify a JSON path to extract from the event and use as the partition key for the Amazon Kinesis data stream, so that you can control the shard to which the event goes. If you do not include this parameter, the default is to use the eventId as the partition key.

ListRuleNamesByTargetRequest

ListRuleNamesByTargetResponse

ListRulesRequest

ListRulesResponse

ListTagsForResourceRequest

ListTagsForResourceResponse

ListTargetsByRuleRequest

ListTargetsByRuleResponse

NetworkConfiguration
This structure specifies the network configuration for an ECS task.

PutEventsRequest

PutEventsRequestEntry
Represents an event to be submitted.

PutEventsResponse

PutEventsResultEntry
Represents an event that failed to be submitted.

PutPermissionRequest

PutRuleRequest

PutRuleResponse

PutTargetsRequest

PutTargetsResponse

PutTargetsResultEntry
Represents a target that failed to be added to a rule.

RemovePermissionRequest

RemoveTargetsRequest

RemoveTargetsResponse

RemoveTargetsResultEntry
Represents a target that failed to be removed from a rule.

Rule
Contains information about a rule in Amazon CloudWatch Events.

RunCommandParameters
This parameter contains the criteria (either InstanceIds or a tag) used to specify which EC2 instances are to be sent the command.

RunCommandTarget
Information about the EC2 instances that are to be sent the command, specified as key-value pairs. Each RunCommandTarget block can include only one key, but this key may specify multiple values.

SqsParameters
This structure includes the custom parameter to be used when the target is an SQS FIFO queue.

Tag
A key-value pair associated with an AWS resource. In CloudWatch Events, rules support tagging.

TagResourceRequest

TagResourceResponse

Target
Targets are the resources to be invoked when a rule is triggered. For a complete list of services and resources that can be set as a target, see PutTargets.

If you are setting the event bus of another account as the target, and that account granted permission to your account through an organization instead of directly by the account ID, then you must specify a RoleArn with proper permissions in the Target structure. For more information, see Sending and Receiving Events Between AWS Accounts in the Amazon CloudWatch Events User Guide.

TestEventPatternRequest

TestEventPatternResponse

UntagResourceRequest

UntagResourceResponse

Enums

DeleteRuleError
Errors returned by DeleteRule

DescribeEventBusError
Errors returned by DescribeEventBus

DescribeRuleError
Errors returned by DescribeRule

DisableRuleError
Errors returned by DisableRule

EnableRuleError
Errors returned by EnableRule

ListRuleNamesByTargetError
Errors returned by ListRuleNamesByTarget

ListRulesError
Errors returned by ListRules

ListTagsForResourceError
Errors returned by ListTagsForResource

ListTargetsByRuleError
Errors returned by ListTargetsByRule

PutEventsError
Errors returned by PutEvents

PutPermissionError
Errors returned by PutPermission

PutRuleError
Errors returned by PutRule

PutTargetsError
Errors returned by PutTargets

RemovePermissionError
Errors returned by RemovePermission

RemoveTargetsError
Errors returned by RemoveTargets

TagResourceError
Errors returned by TagResource

TestEventPatternError
Errors returned by TestEventPattern

UntagResourceError
Errors returned by UntagResource

Traits

CloudWatchEvents
Trait representing the capabilities of the Amazon CloudWatch Events API. Amazon CloudWatch Events clients implement this trait.