Crate rusoto_events

Amazon CloudWatch Events

If you're using the service, you're probably looking for CloudWatchEventsClient and CloudWatchEvents.

Structs

CloudWatchEventsClient
A client for the Amazon CloudWatch Events API.

DeleteRuleRequest

DescribeRuleRequest

DescribeRuleResponse

DisableRuleRequest

EcsParameters
The custom parameters to be used when the target is an Amazon ECS cluster.

EnableRuleRequest

InputTransformer
Contains the parameters needed for you to provide custom input to a target based on one or more pieces of data extracted from the event.

KinesisParameters
This object enables you to specify a JSON path to extract from the event and use as the partition key for the Amazon Kinesis stream, so that you can control the shard to which the event goes. If you do not include this parameter, the default is to use the eventId as the partition key.

ListRuleNamesByTargetRequest

ListRuleNamesByTargetResponse

ListRulesRequest

ListRulesResponse

ListTargetsByRuleRequest

ListTargetsByRuleResponse

PutEventsRequest

PutEventsRequestEntry
Represents an event to be submitted.

PutEventsResponse

PutEventsResultEntry
Represents an event that failed to be submitted.

PutRuleRequest

PutRuleResponse

PutTargetsRequest

PutTargetsResponse

PutTargetsResultEntry
Represents a target that failed to be added to a rule.

RemoveTargetsRequest

RemoveTargetsResponse

RemoveTargetsResultEntry
Represents a target that failed to be removed from a rule.

Rule
Contains information about a rule in Amazon CloudWatch Events.

RunCommandParameters
This parameter contains the criteria (either InstanceIds or a tag) used to specify which EC2 instances are to be sent the command.

RunCommandTarget
Information about the EC2 instances that are to be sent the command, specified as key-value pairs. Each RunCommandTarget block can include only one key, but this key may specify multiple values.

Target
Targets are the resources to be invoked when a rule is triggered. Target types include EC2 instances, AWS Lambda functions, Amazon Kinesis streams, Amazon ECS tasks, AWS Step Functions state machines, Run Command, and built-in targets.

TestEventPatternRequest

TestEventPatternResponse

Enums

DeleteRuleError
Errors returned by DeleteRule

DescribeRuleError
Errors returned by DescribeRule

DisableRuleError
Errors returned by DisableRule

EnableRuleError
Errors returned by EnableRule

ListRuleNamesByTargetError
Errors returned by ListRuleNamesByTarget

ListRulesError
Errors returned by ListRules

ListTargetsByRuleError
Errors returned by ListTargetsByRule

PutEventsError
Errors returned by PutEvents

PutRuleError
Errors returned by PutRule

PutTargetsError
Errors returned by PutTargets

RemoveTargetsError
Errors returned by RemoveTargets

TestEventPatternError
Errors returned by TestEventPattern

Traits

CloudWatchEvents
Trait representing the capabilities of the Amazon CloudWatch Events API. Amazon CloudWatch Events clients implement this trait.

Type Definitions

Arn

Boolean

ErrorCode

ErrorMessage

EventId

EventPattern

EventResource

EventResourceList

EventTime

InputTransformerPathKey

Integer

LimitMax100

LimitMin1

NextToken

PutEventsRequestEntryList

PutEventsResultEntryList

PutTargetsResultEntryList

RemoveTargetsResultEntryList

RoleArn

RuleArn

RuleDescription

RuleName

RuleNameList

RuleResponseList

RuleState

RunCommandTargetKey

RunCommandTargetValue

RunCommandTargetValues

RunCommandTargets

ScheduleExpression

TargetArn

TargetId

TargetIdList

TargetInput

TargetInputPath

TargetList

TargetPartitionKeyPath

TransformerInput

TransformerPaths