Trait rusoto_events::CloudWatchEvents

pub trait CloudWatchEvents {
    fn delete_rule(
        &self, 
        input: DeleteRuleRequest
    ) -> RusotoFuture<(), DeleteRuleError>;

    fn describe_event_bus(
        &self
    ) -> RusotoFuture<DescribeEventBusResponse, DescribeEventBusError>;

    fn describe_rule(
        &self, 
        input: DescribeRuleRequest
    ) -> RusotoFuture<DescribeRuleResponse, DescribeRuleError>;

    fn disable_rule(
        &self, 
        input: DisableRuleRequest
    ) -> RusotoFuture<(), DisableRuleError>;

    fn enable_rule(
        &self, 
        input: EnableRuleRequest
    ) -> RusotoFuture<(), EnableRuleError>;

    fn list_rule_names_by_target(
        &self, 
        input: ListRuleNamesByTargetRequest
    ) -> RusotoFuture<ListRuleNamesByTargetResponse, ListRuleNamesByTargetError>;

    fn list_rules(
        &self, 
        input: ListRulesRequest
    ) -> RusotoFuture<ListRulesResponse, ListRulesError>;

    fn list_targets_by_rule(
        &self, 
        input: ListTargetsByRuleRequest
    ) -> RusotoFuture<ListTargetsByRuleResponse, ListTargetsByRuleError>;

    fn put_events(
        &self, 
        input: PutEventsRequest
    ) -> RusotoFuture<PutEventsResponse, PutEventsError>;

    fn put_permission(
        &self, 
        input: PutPermissionRequest
    ) -> RusotoFuture<(), PutPermissionError>;

    fn put_rule(
        &self, 
        input: PutRuleRequest
    ) -> RusotoFuture<PutRuleResponse, PutRuleError>;

    fn put_targets(
        &self, 
        input: PutTargetsRequest
    ) -> RusotoFuture<PutTargetsResponse, PutTargetsError>;

    fn remove_permission(
        &self, 
        input: RemovePermissionRequest
    ) -> RusotoFuture<(), RemovePermissionError>;

    fn remove_targets(
        &self, 
        input: RemoveTargetsRequest
    ) -> RusotoFuture<RemoveTargetsResponse, RemoveTargetsError>;

    fn test_event_pattern(
        &self, 
        input: TestEventPatternRequest
    ) -> RusotoFuture<TestEventPatternResponse, TestEventPatternError>;
}

Trait representing the capabilities of the Amazon CloudWatch Events API. Amazon CloudWatch Events clients implement this trait.

Required Methods

fn delete_rule(
&self,
input: DeleteRuleRequest
) -> RusotoFuture<(), DeleteRuleError>

Deletes the specified rule.

You must remove all targets from a rule using RemoveTargets before you can delete the rule.

When you delete a rule, incoming events might continue to match to the deleted rule. Please allow a short period of time for changes to take effect.

fn describe_event_bus(
&self
) -> RusotoFuture<DescribeEventBusResponse, DescribeEventBusError>

Displays the external AWS accounts that are permitted to write events to your account using your account's event bus, and the associated policy. To enable your account to receive events from other accounts, use PutPermission.

fn describe_rule(
&self,
input: DescribeRuleRequest
) -> RusotoFuture<DescribeRuleResponse, DescribeRuleError>

Describes the specified rule.

fn disable_rule(
&self,
input: DisableRuleRequest
) -> RusotoFuture<(), DisableRuleError>

Disables the specified rule. A disabled rule won't match any events, and won't self-trigger if it has a schedule expression.

When you disable a rule, incoming events might continue to match to the disabled rule. Please allow a short period of time for changes to take effect.

fn enable_rule(
&self,
input: EnableRuleRequest
) -> RusotoFuture<(), EnableRuleError>

Enables the specified rule. If the rule does not exist, the operation fails.

When you enable a rule, incoming events might not immediately start matching to a newly enabled rule. Please allow a short period of time for changes to take effect.

fn list_rule_names_by_target(
&self,
input: ListRuleNamesByTargetRequest
) -> RusotoFuture<ListRuleNamesByTargetResponse, ListRuleNamesByTargetError>

Lists the rules for the specified target. You can see which of the rules in Amazon CloudWatch Events can invoke a specific target in your account.

fn list_rules(
&self,
input: ListRulesRequest
) -> RusotoFuture<ListRulesResponse, ListRulesError>

Lists your Amazon CloudWatch Events rules. You can either list all the rules or you can provide a prefix to match to the rule names.

fn list_targets_by_rule(
&self,
input: ListTargetsByRuleRequest
) -> RusotoFuture<ListTargetsByRuleResponse, ListTargetsByRuleError>

Lists the targets assigned to the specified rule.

fn put_events(
&self,
input: PutEventsRequest
) -> RusotoFuture<PutEventsResponse, PutEventsError>

Sends custom events to Amazon CloudWatch Events so that they can be matched to rules.

fn put_permission(
&self,
input: PutPermissionRequest
) -> RusotoFuture<(), PutPermissionError>

Running PutPermission permits the specified AWS account to put events to your account's default event bus. CloudWatch Events rules in your account are triggered by these events arriving to your default event bus.

For another account to send events to your account, that external account must have a CloudWatch Events rule with your account's default event bus as a target.

To enable multiple AWS accounts to put events to your default event bus, run PutPermission once for each of these accounts.

The permission policy on the default event bus cannot exceed 10KB in size.

fn put_rule(
&self,
input: PutRuleRequest
) -> RusotoFuture<PutRuleResponse, PutRuleError>

Creates or updates the specified rule. Rules are enabled by default, or based on value of the state. You can disable a rule using DisableRule.

If you are updating an existing rule, the rule is completely replaced with what you specify in this PutRule command. If you omit arguments in PutRule, the old values for those arguments are not kept. Instead, they are replaced with null values.

When you create or update a rule, incoming events might not immediately start matching to new or updated rules. Please allow a short period of time for changes to take effect.

A rule must contain at least an EventPattern or ScheduleExpression. Rules with EventPatterns are triggered when a matching event is observed. Rules with ScheduleExpressions self-trigger based on the given schedule. A rule can have both an EventPattern and a ScheduleExpression, in which case the rule triggers on matching events as well as on a schedule.

Most services in AWS treat : or / as the same character in Amazon Resource Names (ARNs). However, CloudWatch Events uses an exact match in event patterns and rules. Be sure to use the correct ARN characters when creating event patterns so that they match the ARN syntax in the event you want to match.

fn put_targets(
&self,
input: PutTargetsRequest
) -> RusotoFuture<PutTargetsResponse, PutTargetsError>

Adds the specified targets to the specified rule, or updates the targets if they are already associated with the rule.

Targets are the resources that are invoked when a rule is triggered.

You can configure the following as targets for CloudWatch Events:

EC2 instances

AWS Lambda functions

Streams in Amazon Kinesis Streams

Delivery streams in Amazon Kinesis Firehose

Amazon ECS tasks

AWS Step Functions state machines

AWS Batch jobs

Pipelines in Amazon Code Pipeline

Amazon Inspector assessment templates

Amazon SNS topics

Amazon SQS queues, including FIFO queues

The default event bus of another AWS account

Note that creating rules with built-in targets is supported only in the AWS Management Console.

For some target types, PutTargets provides target-specific parameters. If the target is an Amazon Kinesis stream, you can optionally specify which shard the event goes to by using the KinesisParameters argument. To invoke a command on multiple EC2 instances with one rule, you can use the RunCommandParameters field.

To be able to make API calls against the resources that you own, Amazon CloudWatch Events needs the appropriate permissions. For AWS Lambda and Amazon SNS resources, CloudWatch Events relies on resource-based policies. For EC2 instances, Amazon Kinesis streams, and AWS Step Functions state machines, CloudWatch Events relies on IAM roles that you specify in the RoleARN argument in PutTargets. For more information, see Authentication and Access Control in the Amazon CloudWatch Events User Guide.

If another AWS account is in the same region and has granted you permission (using PutPermission), you can send events to that account by setting that account's event bus as a target of the rules in your account. To send the matched events to the other account, specify that account's event bus as the Arn when you run PutTargets. If your account sends events to another account, your account is charged for each sent event. Each event sent to antoher account is charged as a custom event. The account receiving the event is not charged. For more information on pricing, see Amazon CloudWatch Pricing.

For more information about enabling cross-account events, see PutPermission.

Input, InputPath and InputTransformer are mutually exclusive and optional parameters of a target. When a rule is triggered due to a matched event:

If none of the following arguments are specified for a target, then the entire event is passed to the target in JSON form (unless the target is Amazon EC2 Run Command or Amazon ECS task, in which case nothing from the event is passed to the target).

If Input is specified in the form of valid JSON, then the matched event is overridden with this constant.

If InputPath is specified in the form of JSONPath (for example, $.detail), then only the part of the event specified in the path is passed to the target (for example, only the detail part of the event is passed).

If InputTransformer is specified, then one or more specified JSONPaths are extracted from the event and used as values in a template that you specify as the input to the target.

When you specify InputPath or InputTransformer, you must use JSON dot notation, not bracket notation.

When you add targets to a rule and the associated rule triggers soon after, new or updated targets might not be immediately invoked. Please allow a short period of time for changes to take effect.

This action can partially fail if too many requests are made at the same time. If that happens, FailedEntryCount is non-zero in the response and each entry in FailedEntries provides the ID of the failed target and the error code.

fn remove_permission(
&self,
input: RemovePermissionRequest
) -> RusotoFuture<(), RemovePermissionError>

Revokes the permission of another AWS account to be able to put events to your default event bus. Specify the account to revoke by the StatementId value that you associated with the account when you granted it permission with PutPermission. You can find the StatementId by using DescribeEventBus.

fn remove_targets(
&self,
input: RemoveTargetsRequest
) -> RusotoFuture<RemoveTargetsResponse, RemoveTargetsError>

Removes the specified targets from the specified rule. When the rule is triggered, those targets are no longer be invoked.

When you remove a target, when the associated rule triggers, removed targets might continue to be invoked. Please allow a short period of time for changes to take effect.

This action can partially fail if too many requests are made at the same time. If that happens, FailedEntryCount is non-zero in the response and each entry in FailedEntries provides the ID of the failed target and the error code.

fn test_event_pattern(
&self,
input: TestEventPatternRequest
) -> RusotoFuture<TestEventPatternResponse, TestEventPatternError>

Tests whether the specified event pattern matches the provided event.

Most services in AWS treat : or / as the same character in Amazon Resource Names (ARNs). However, CloudWatch Events uses an exact match in event patterns and rules. Be sure to use the correct ARN characters when creating event patterns so that they match the ARN syntax in the event you want to match.